backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
opensearch-py==2.3.1
python-dotenv==1.0.0
